Refactoring: Move all Fedora specific tests into test_install_fedora.py
Adding helpers and fixtures to detect Fedora based
